requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
firebase-admin
pyjwt
